修正

+++ Android/Android1/Android1.NativeActivity/Apple.cpp
@@ -55,8 +55,7 @@ void Apple::UpData(void)
 
 	if (effect_cnt > 0)
 	{
-		pos.y += cos(RAD(effect_cnt*6))*15;
-		//pos.y++;
+		pos.y += (int)(cosf(RAD(effect_cnt * 6)) * 15);
 		effect_cnt--;
 	}
 
@@ -71,9 +70,7 @@ void Apple::UpData(void)
 			if (CheckHit(center, { 1,1 }, pl.lock()->GetLocalPos(), st.lock()->GetChipPlSize()))
 			{
 				Sound::Get()->Play(SE_HEAL);
-
 			}
-
 		}
 	}
